fix some todos, particularly authentication related ones

+++ src/main/java/com/samebug/clients/common/search/api/client/Config.java
@@ -21,7 +21,7 @@ public class Config {
     @Nullable
     public String apiKey;
     @Nullable
-    public Long workspaceId;
+    public Integer workspaceId;
     public String serverRoot;
     public String trackingRoot;
     public boolean isTrackingEnabled;
